Fix all route lints
