--- api/lib/basemap.ts
@@ -1,3 +1,0 @@
-export default class BasemapManager {
-
-}
