+++ api/routes/connection-channels.ts
@@ -3,8 +3,7 @@ import Auth, { AuthResourceAccess } from '../lib/auth.js';
 import Config from '../lib/config.js';
 import TAKAPI, { APIAuthCertificate, } from '../lib/tak-api.js';
 import { Type } from '@sinclair/typebox'
-import { GenericListOrder } from '@openaddresses/batch-generic';
-import { StandardResponse, ConnectionResponse, GenericMartiResponse } from '../lib/types.js';
+import { GenericMartiResponse } from '../lib/types.js';
 import Schema from '@openaddresses/batch-schema';
 
 export default async function router(schema: Schema, config: Config) {
