+++ api/routes/connection-token.ts
@@ -1,9 +1,7 @@
 import { Type } from '@sinclair/typebox'
-import { Response } from 'express';
 import Err from '@openaddresses/batch-error';
 import Auth from '../lib/auth.js';
 import Config from '../lib/config.js';
-import { promisify } from 'util';
 import jwt from 'jsonwebtoken';
 import { sql } from 'drizzle-orm';
 import { GenericListOrder } from '@openaddresses/batch-generic';
