+++ api/routes/layer.ts
@@ -2,33 +2,23 @@ import { Type, Static } from '@sinclair/typebox'
 import sleep from '../lib/sleep.js';
 import { GenericListOrder } from '@openaddresses/batch-generic';
 import Schema from '@openaddresses/batch-schema';
-import { check } from '@placemarkio/check-geojson';
-import bodyparser from 'body-parser';
 import Err from '@openaddresses/batch-error';
-import { CoT } from '@tak-ps/node-tak';
-import { Item as QueueItem } from '../lib/queue.js'
 import Cacher from '../lib/cacher.js';
 import Auth, { AuthResourceAccess } from '../lib/auth.js';
 import Lambda from '../lib/aws/lambda.js';
 import CloudFormation from '../lib/aws/cloudformation.js';
 import Style, { StyleContainer } from '../lib/style.js';
 import Alarm from '../lib/aws/alarm.js';
-import DDBQueue from '../lib/queue.js';
 import Config from '../lib/config.js';
 import Schedule from '../lib/schedule.js';
-import S3 from '../lib/aws/s3.js';
-import { Feature } from 'geojson';
 import { Param } from '@openaddresses/batch-generic';
 import { sql } from 'drizzle-orm';
 import { StandardResponse, LayerResponse, Layer_Config } from '../lib/types.js';
 import { Layer_Priority } from '../lib/enums.js';
 import { Layer } from '../lib/schema.js';
-import TAKAPI, { APIAuthCertificate, } from '../lib/tak-api.js';
 
 export default async function router(schema: Schema, config: Config) {
     const alarm = new Alarm(config.StackName);
-    const ddb = new DDBQueue(config.StackName);
-    ddb.on('error', (err) => { console.error(err); });
 
     await schema.get('/layer', {
         name: 'List Layers',
@@ -146,7 +136,7 @@ export default async function router(schema: Schema, config: Config) {
             }
 
             Schedule.is_valid(req.body.cron);
-            let layer = await config.models.Layer.generate(req.body);
+            const layer = await config.models.Layer.generate(req.body);
 
             if (config.events) {
                 if (layer.cron && !Schedule.is_aws(layer.cron) && layer.enabled) {
@@ -409,129 +399,6 @@ export default async function router(schema: Schema, config: Config) {
             return res.json({
                 status: 200,
                 message: 'Layer Deleted'
-            });
-        } catch (err) {
-            return Err.respond(err, res);
-        }
-    });
-
-    await schema.post('/layer/:layerid/cot', {
-        name: 'Post COT',
-        group: 'Layer',
-        description: 'Post CoT data to a given layer',
-        params: Type.Object({
-            layerid: Type.Integer()
-        }),
-        query: Type.Object({
-            logging: Type.Optional(Type.Boolean({ "description": "If logging is enabled for the layer, allow callers to skip logging for a particular CoT payload" }))
-        }),
-        body: Type.Any(),
-        res: StandardResponse
-    }, async (req, res) => {
-        try {
-            await Auth.is_auth(config, req, {
-                resources: [{ access: AuthResourceAccess.LAYER, id: req.params.layerid }]
-            });
-
-            if (!req.headers['content-type']) throw new Err(400, null, 'Content-Type not set');
-
-            const layer = await config.cacher.get(Cacher.Miss(req.query, `layer-${req.params.layerid}`), async () => {
-                return await config.models.Layer.from(req.params.layerid);
-            });
-
-            const style = new Style(layer);
-
-            try {
-                req.body = check(JSON.stringify(req.body));
-            } catch (err) {
-                throw new Err(400, null, err instanceof Error ? err.message : String(err));
-            }
-
-            for (let i = 0; i < req.body.features.length; i++) {
-                req.body.features[i] = await style.feat(req.body.features[i])
-            }
-
-            let pooledClient;
-            let data;
-            const cots = [];
-            if (!layer.data && layer.connection) {
-                pooledClient = await config.conns.get(layer.connection);
-
-                for (const feat of req.body.features) {
-                    cots.push(CoT.from_geojson(feat))
-                }
-            } else if (layer.data) {
-                data = await config.models.Data.from(layer.data);
-
-                if (!data.mission_sync) {
-                    return res.status(202).json({ status: 202, message: 'Recieved but Data Mission Sync Disabled' });
-                }
-
-                pooledClient = await config.conns.get(data.connection);
-
-                if (data.mission_diff) {
-                    if (!Array.isArray(req.body.uids)) throw new Err(400, null, 'uids Array must be present when submitting to DataSync with MissionDiff');
-
-                    const api = await TAKAPI.init(new URL(String(config.server.api)), new APIAuthCertificate(pooledClient.config.auth.cert, pooledClient.config.auth.key));
-                    // Once NodeJS supports Set.difference we can simplify this
-                    const inputFeats = new Set(req.body.uids);
-                    const features = await api.Mission.latestFeats(data.name, { token: data.mission_token });
-
-                    for (const feat of features) {
-                        if (!inputFeats.has(String(feat.id))) {
-                            await api.Mission.detachContents(
-                                data.name,
-                                { uid: String(feat.id) },
-                                { token: data.mission_token }
-                            );
-                        }
-                    }
-                }
-
-                for (const feat of req.body.features) {
-                    feat.properties.dest = [{ mission: data.name }];
-                    cots.push(CoT.from_geojson(feat))
-                }
-            } else {
-                throw new Err(400, null, 'Either connection or data must be set');
-            }
-
-            if (!pooledClient || !pooledClient.config || !pooledClient.config.enabled) {
-                return res.json({ status: 200, message: 'Recieved but Connection Paused' });
-            }
-
-            if (cots.length === 0) return res.json({ status: 200, message: 'No features found' });
-
-            pooledClient.tak.write(cots);
-            for (const cot of cots) config.conns.cot(pooledClient.config, cot);
-
-            if (data) {
-                const api = await TAKAPI.init(new URL(String(config.server.api)), new APIAuthCertificate(pooledClient.config.auth.cert, pooledClient.config.auth.key));
-                const missionContent = await api.Mission.attachContents(data.name, {
-                    uids: cots.map((cot) => { return cot.raw.event._attributes.uid })
-                }, {
-                    token: data.mission_token
-                });
-            }
-
-            // TODO Only GeoJSON Features go to Dynamo, this should also store CoT XML
-            // @ts-ignore
-            if (layer.logging && req.query.logging !== false) ddb.queue(req.body.features.map((feat: Feature) => {
-                const item: QueueItem = {
-                    id: String(feat.id),
-                    layer: layer.id,
-                    type: feat.type,
-                    // @ts-ignore
-                    properties: feat.properties,
-                    geometry: feat.geometry
-                }
-
-                return item;
-            }));
-
-            res.json({
-                status: 200,
-                message: 'Submitted'
             });
         } catch (err) {
             return Err.respond(err, res);
