+++ api/routes/marti-files.ts
@@ -5,9 +5,7 @@ import Err from '@openaddresses/batch-error';
 import Auth from '../lib/auth.js';
 import Config from '../lib/config.js';
 import TAKAPI, {
-    APIAuthToken,
     APIAuthCertificate,
-    APIAuthPassword
 } from '../lib/tak-api.js';
 
 export default async function router(schema: Schema, config: Config) {
@@ -27,7 +25,7 @@ export default async function router(schema: Schema, config: Config) {
             const user = await Auth.as_user(config, req, { token: true });
             const profile = await config.models.Profile.from(user.email);
             const api = await TAKAPI.init(new URL(String(config.server.api)), new APIAuthCertificate(profile.auth.cert, profile.auth.key));
-            const file = await api.Files.delete(req.params.hash);
+            await api.Files.delete(req.params.hash);
 
             return res.json({
                 status: 200,
