+++ api/routes/marti-package.ts
@@ -1,21 +1,14 @@
-import fs from 'node:fs';
 import crypto from 'node:crypto';
 import { Type } from '@sinclair/typebox'
 import CoT, { FileShare, DataPackage } from '@tak-ps/node-cot';
-import TAK from '@tak-ps/node-tak';
 import Schema from '@openaddresses/batch-schema';
 import Err from '@openaddresses/batch-error';
 import Auth from '../lib/auth.js';
 import Config from '../lib/config.js';
-import { StandardResponse, GenericMartiResponse } from '../lib/types.js';
 import { Content } from '../lib/api/files.js';
 import { Package } from '../lib/api/package.js';
-import { Profile } from '../lib/schema.js';
-import S3 from '../lib/aws/s3.js';
 import TAKAPI, {
-    APIAuthToken,
     APIAuthCertificate,
-    APIAuthPassword
 } from '../lib/tak-api.js';
 
 export default async function router(schema: Schema, config: Config) {
