+++ api/routes/profile.ts
@@ -1,7 +1,7 @@
 import { Type } from '@sinclair/typebox'
 import Schema from '@openaddresses/batch-schema';
 import Err from '@openaddresses/batch-error';
-import Auth, { AuthResource } from '../lib/auth.js';
+import Auth from '../lib/auth.js';
 import { ProfileResponse } from '../lib/types.js'
 import Config from '../lib/config.js';
 import { TAKRole, TAKGroup } from '../lib/api/types.js'
@@ -36,10 +36,10 @@ export default async function router(schema: Schema, config: Config) {
             tak_callsign: Type.Optional(Type.String()),
             tak_group: Type.Optional(Type.Enum(TAKGroup)),
             tak_role: Type.Optional(Type.Enum(TAKRole)),
-            tak_loc: Type.Optional([Type.Null(), Type.Object({
+            tak_loc: Type.Optional(Type.Union([Type.Null(), Type.Object({
                 type: Type.String(),
                 coordinates: Type.Array(Type.Number())
-            })])
+            })]))
         }),
         res: ProfileResponse
     }, async (req, res) => {
