+++ api/routes/search.ts
@@ -21,7 +21,7 @@ export default async function router(schema: Schema, config: Config) {
         })
     }, async (req, res) => {
         try {
-            const user = await Auth.as_user(config, req);
+            await Auth.as_user(config, req);
             const response = {
                 weather: null
             };
